+++ bld/ui/h/uidialog.h
@@ -2,7 +2,7 @@
 *
 *                            Open Watcom Project
 *
-* Copyright (c) 2002-2020 The Open Watcom Contributors. All Rights Reserved.
+* Copyright (c) 2002-2021 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -68,7 +68,7 @@ typedef bool (UIPICKGETTEXT)( const void *data_handle, unsigned item, char *buff
 
 typedef struct an_edit_control {
     char            *buffer;
-    unsigned        length;
+    size_t          length;
 } an_edit_control;
 
 typedef struct a_list_info {
